Replace `failure` with `anyhow`/`thiserror` (#24)
`failure` is [deprecated](https://github.com/rust-lang-deprecated/failure#failure---a-new-error-management-story). `anyhow` is a recommended replacement for `failure::Error`, including backtraces (with the `backtrace` feature).

## Details

`read_pair_iter::FastqError` is not a `failure::Fail` anymore, it is an `std::error::Error` derived with `thiserror`. I had to remove the `backtrace` field because `thiserror` is based on `std::backtrace`, which is [not stable yet](https://github.com/dtolnay/thiserror#details). This is compensated by `anyhow::Error` being used in public interfaces, which [guarantees a backtrace is available](https://docs.rs/anyhow/1.0.44/anyhow/struct.Error.html#).